+++ R/bb_options.R
@@ -45,6 +45,12 @@
 #'   \item{`sal_max_jump`}{The maximum difference between successive salinity
 #'   records before the salinity jump (`Sj`) flag is set for both involved
 #'   records.}
+#'   \item{`min_depth`, `max_depth`}{ The minimum and maximum water depth. If
+#'   the `Depth` column is present and falls outside of this range either the
+#'   `Wl` (water depth low) or `Wh` (water depth high) flags will be thrown.
+#'   Additionaly `7` will be written to `Depth_QC` and either `9999`  or
+#'   `7` will be written to `Gen_QC` depending on if it's high or low
+#'   respectively}
 #'
 #'}
 #'